mockmate2/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
eventlet==0.33.3
PyPDF2==3.0.1
pyttsx3==2.90
openai-whisper==20230314
SpeechRecognition==3.10.0
python-socketio==5.8.0